Mail: Better placement of signature menu
When clicking the "Signature" button in the toolbar, no mouse
coordinates are available. The current hack is to open the menu
in the center of the tool bar, which is normally no way near
where the user has clicked the button. Not nice.

Change that to find the button coordinates and place the menu a bit
to the left of its center (so chances are the menu pops up already
under the mouse).

Fall back to the old hack, if the signature button wasn't found for
some reason...

Change-Id: I8fb5dbe4e2d394768a685fda0dc71d3cb2c5961d
Reviewed-on: https://review.haiku-os.org/c/haiku/+/1836
Reviewed-by: Adrien Destugues <pulkomandy@gmail.com>