kernel/disk_device_manager: Continue scanning after invalid partitions.
As the inline comment notes, just because we could not scan one
partition does not mean we won't be able to scan the following ones.
This fixes scanning for disks on certain systems following axeld's
change in 7c2135591ba90958b5c8d92a2a5ef8a930c9ee28.

Part of #15330.