ext2: add test script for ext2 filesystem.
based on a test script from FreeBSD.
this uses fsx and fstorture to exercise the fs, mkfs.ext3 to initialize,
e2fsck to check the integrity. single-threaded test is OK, multi-threaded is not.
the following blocksizes and features are tested: 1K, 2K and 4K; huge_file,
dir_nlink, extent, 64bit, metadata_csum.

Change-Id: I948672194bfae0132341183efa4f4af962e2de02
Reviewed-on: https://review.haiku-os.org/c/haiku/+/1843
Reviewed-by: Rene Gollent <rene@gollent.com>